requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.30.0
authx>=1.4.0
passlib[bcrypt]>=1.7.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
email-validator>=2.0.0
pydantic-settings>=2.1.0
aiofiles>=24.1.0 